correct typo

+++ mia/core/mitestimages.hh
@@ -39,8 +39,8 @@ struct SImageSize {
 }; 
 
 extern EXPORT_CORE const SImageSize mi_test_size; 
-extern EXPORT_CORE const int moving_init_data[]; 
-extern EXPORT_CORE const int reference_init_data[]; 
+extern EXPORT_CORE const float moving_init_data[]; 
+extern EXPORT_CORE const float reference_init_data[]; 
 
 NS_MIA_END
